requirements.txt
Flask==2.0.1 
Werkzeug==2.0.2 
requests==2.26.0
python-decouple==3.5

